Refactored out background card and progress bar
